+++ test/libsolidity/syntaxTests/constants/access_through_contract_name.sol
@@ -1,0 +1,8 @@
+contract A {
+    uint constant a = 0x1;
+}
+contract C {
+    function f() public pure returns (uint) {
+        return A.a;
+    }
+}