+++ test/libsolidity/syntaxTests/constants/cross_contract_cyclic_dependency.sol
@@ -1,0 +1,8 @@
+contract A {
+    uint constant c = B.c;
+}
+contract B {
+    uint constant c = A.c;
+}
+// ----
+// Some Error: This should fail with some error!